Cleanup

+++ device/api/umd/device/chip_helpers/tlb_manager.h
@@ -47,7 +47,7 @@ public:
     static const std::vector<uint32_t> get_tlb_arch_sizes(const tt::ARCH arch);
 
     std::unique_ptr<TlbWindow> allocate_tlb_window(
-        tlb_data config, const tt::umd::TlbMapping mapping = tt::umd::TlbMapping::WC, const uint32_t tlb_size = 0);
+        tlb_data config, const tt::umd::TlbMapping mapping = tt::umd::TlbMapping::WC, const size_t tlb_size = 0);
 
 private:
     TTDevice* tt_device_;
